Add fuzzing and fix crashers.
The output of strings.ToLower can have less bytes than the input.
There are valid utf8 strings which demonstrate this which could be
entered as potential passwords which would cause an out of bounds
slice access in the dictionary match code.

Fixed issue in go 1.10 test code for adjcmatrix.go to make sure
the tests passed cleanly.